+++ zerver/views/auth.py
@@ -54,6 +54,7 @@ from zerver.lib.exceptions import (
     PasswordResetRequiredError,
     RateLimited,
     RealmDeactivatedError,
+    RealmMovedError,
     UserDeactivatedError,
 )
 from zerver.lib.mobile_auth_otp import otp_encrypt_api_key
@@ -936,6 +937,14 @@ def check_server_incompatibility(request: HttpRequest) -> bool:
 @csrf_exempt
 def api_get_server_settings(request: HttpRequest) -> HttpResponse:
     # Log which client is making this request.
+
+    realm = get_realm_from_request(request)
+    if realm is not None and realm.deactivated:
+        if realm.deactivated_redirect:
+            raise RealmMovedError(realm.deactivated_redirect)
+        else:
+            raise RealmDeactivatedError()
+
     process_client(request, request.user, skip_update_user_activity=True)
     result = dict(
         authentication_methods=get_auth_backends_data(request),