+++ templates/zerver/api/changelog.md
@@ -20,6 +20,11 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 5.0
 
+**Feature level 113**
+
+* [`GET /server_settings`](/api/get-server-settings): The endpoint
+  now returns an error when requesting a deactivated realm.
+
 **Feature level 112**
 
 * [`GET /events`](/api/get-events): Updated `update_message` event type